Fix the form submission when exceptions are raised while updating values
When Set_Value is called, the Ada bean can raise an exception.  That exception
is catched and handled and put in the exception queue but we must not
continue the form submission and the form action must not be called.
When this happens, we must render the response with the error message
(the error message was assigned to the correct field)